Don't turn on/off screen during suspend/resume.
Currently we turn off all the displays during suspend and turn back
on all the displays during resume. This adds noticeable delay to
the resume time. So this patch gets rid of turning screen on/off
during suspend/resume and relies on the kernel driver to correctly
save/restore the display states.

One exception is that if system goes from IDLE_OFF to SUSPENDED, then
when resuming, kernel driver will bring the screen back to IDLE_OFF
state, which is blank screen. User probably won't expect that. So if
the state before SUSPENDED is IDLE_OFF, we explicitly turn the screen
on when resuming.

BUG=None
TEST=1.Log in as guest or user, close the lid to suspend, open the
       lid to resume and make sure there is no significant amount
       of time during which the screen is blank.
     2.Connect chromebook with external monitors, go into suspend mode,
       by closing the lid or idle off, resume the device and make sure
       screen comes back to both LCD and external monitor.

Change-Id: Ia2eb032cd8a7995b9e6d89a0c24818516a772462
Reviewed-on: https://gerrit.chromium.org/gerrit/18918
Commit-Ready: Yufeng Shen <miletus@chromium.org>
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>