shill: Allow creation and push of DefaultProfile
Add the ability to create a DefaultProfile whose name is not
"default".  Add the ability to push a DefaultProfile atop
another DefaultProfile in the Manager.  This will enable
autotests to run that can test interactions between profiles
while a user is not logged in.

BUG=chromium-os:24461
TEST=New unit tests; autotests to come

Change-Id: Iccfcff6ec613cf0a8ac55c11c9cc1efac8237807
Reviewed-on: https://gerrit.chromium.org/gerrit/19287
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>