Add support for passing up if a battery time estimate is suspect
Immediately after a battery state event has occured the charge/discharge rate is
inaccurate. This means that the battery time estimate we calculate is wrong, so
what were previously displaying to the user was plan wrong.

This CL adds support in powerd for flagging the time estimates as suspect. The
UX will check for this and display a calculating string instead of the time
estimate when we don't trust the estimate. Support for this has already been
landed for the protobuf and in Chrome, so all that is needed to turn on the
feature if this CL.

This CL also adds a field to the call to GetPowerStatus, which is a bool that
indicates whether or not we trust the time estimates at this point. This is used
to populate the underlying data stucture and by association protobuf.

BUG=chromium-os:28881
BUG=chromium-os:29225
TEST=Build image and confirm that calculating message appears immediately after
     pulling/plugging power. Then about 5 to 10 seconds after a sane estimate
     for the time should appear.
TEST=Build and run unit tests, to verify new tests pass.

Change-Id: I05f169ea01847187fd1c01ba6a34dd7edd305c22
Reviewed-on: https://gerrit.chromium.org/gerrit/19682
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>