Prevent polling the battery near a power event
This CL disables directly polling when an event about the power state of the
system has been received. It also reschedules the timer based polling from
occuring too near an event. This is needed since the discharge/charge rate
reported by the battery around an event change is inaccurate and should be
used. This an interim fix ment to be back ported to M19. The correct fix for the
issue is being worked on and is intended for M20.

BUG=chromium-os:22988
TEST=Take device off of AC, watch the power status for atleast 30 seconds and
     make sure that when it transitions it doesn't report obviously wrong
     discharge time estimates. Repeat this multiple times to confirm, also watch
     when putting it back on AC for a similar issue with charge time estimates.

Change-Id: Ifc511865852c0f0d01363069192f983b11ce4383
Reviewed-on: https://gerrit.chromium.org/gerrit/19475
Tested-by: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Simon Que <sque@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>