Don't drop cache.
It's possible (probable?) that the sync and cache drop operations part way
through the update process are causing UI jank. However, these operations were
added to work around an apparent kernel bug early in our development. This
bug may or may not have been fixed.

The sync operation here will be re-added during the postinst immediately
before we active the new partition in firmware.

If the kernel bug re-appears, we'll re-add the operations needed to
work around it in postinst so that we have the maximum flexibility.

BUG=chromium-os:27957
TEST=image_to_live.sh.

Change-Id: I152e494c07d3dc9a8e87ea377911d740511c5a30
Reviewed-on: https://gerrit.chromium.org/gerrit/19632
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Commit-Ready: Don Garrett <dgarrett@chromium.org>