tpm: move status files off future encrypted mount
In the case where the TPM is being owned but has not finished, the
future encrypted mount for /var/lib will still not be recoverable.
As a result, the TPM status file should live explicitly in the
unencrypted stateful partition.

This changes the paths and provides migration from the old paths.

BUG=chromium-os:22712
TEST=lumpy build & manual testing

Change-Id: I1811b31709627341aa47a8c3a97c9bf60f8ed4fb
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19930
Reviewed-by: Elly Jones <ellyjones@chromium.org>