Implemented asynchronous initialization and import of tokens.
When a token is initialized there is some expensive TPM work to do. If
legacy objects need to be imported there is even more TPM work to do.
Objects that are not encrypted with the TPM can be available before this
work is complete. This CL enables this immediate access so certificates
can be enumerated before TPM work is finished. The TPM work executes
asynchronously in the background.

To accomplish this the following changes have been made:
- Separate loading of public and private objects.
- Do not encrypt public objects with the TPM.
- Integrate import logic with object loading logic in the object pool.
- Add locking for the handle manager, tpm utility, and object pools.
- Separate importing of public and private objects.

BUG=chromium-os:21014
TEST=Ran all unit tests; manually tested expensive import, ensuring that
UI was not blocked on the completion of the import.

Change-Id: I2620cd2000b52d9d6d547453441a2286018e05b7