Include devices without media as auto-mountable devices.
This CL changes UdevDevice::IsAutoMountable() to include devices without
media such that a DiskAdded event will be sent to the file manager after
a DeviceAdded event for a device without media. The DiskAdded event is
needed by the file manager to dismiss the device notification UI
triggered by the DeviceAdded event. As a result of this change, the file
manager will check to make sure that a device has media before mounting
it.

BUG=chromium-os:27894
TEST=Tested the following:
1. Build cros-disks for {x86,amd64,arm}-generic.
2. Run cros-disks unit tests.
3. Manually insert a device without media to a Cr48 and observe the
   notification from the file manager.

Change-Id: I4524bce0bc8bf9636598c5f11d8134fbdc2d37c0
Reviewed-on: https://gerrit.chromium.org/gerrit/19938
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Ben Chan <benchan@chromium.org>