Add FileToString utility function.
Add a new utility function to read the contents of a file into a string, I need
this in another CL, but thought it made sense to introduce this utility
method first.

Refactor LsbReleaseValue function to use this helper, instead of streams.
Nothing else referred to streams, so this reduces the number of libraries
linked in, and shrinks our stripped executable size by about 300k.

BUG=None
TEST=Added/ran unittests.

Change-Id: I78beaabbc5d6167efa7abede1ce3d3619ede3413
Reviewed-on: https://gerrit.chromium.org/gerrit/20039
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>