Remove flakey test from MetricsStore
Removing the MetricsStoreTests.SyncStoreInvalidMap unit test. This test is
flakey since the memory value it is testing on isn't guaranteed to be aligned
correctly. This test is not really catching the behaviour it is intended to
catch, so it isn't worth fixing. I am going to be rewritting how the metrics
store tracks if memory has been mapped in the near future to fix the meta-issue
this exposed. Thus I am killing the test now, so that it doesn't cause flakey
test failures.

BUG=chromium-os:28319
TEST=Run powerd_unittests

Change-Id: I066917755f38c7972cb9ba965a7acc5f0583bbd3
Reviewed-on: https://gerrit.chromium.org/gerrit/20112
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>