Add WriteStringToFile and ReplaceInFile helper methods.
Add two more helper methods and tests. One writes out to a file from
a string, and the other does a search and replace inside a file
to replace a string.

This later is a replacement for sed -i, the former is just a helper for
the later.

BUG=None
TEST=New unit tests added (and they caught two bugs in original implementation)

Change-Id: I82738067174642d1946861ec2d1475b1b8295a22
Reviewed-on: https://gerrit.chromium.org/gerrit/20291
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>