[cryptohome] move slow credential tests off Mount
Move the slow credential tests (the ones that talk to the TPM) off Mount and
onto HomeDirs, and change existing uses of credential tests to try the fast
tests first and then the slow tests. Note that currently we do the fast test
every mount before trying to slow test, but it doesn't matter much - the
test requires no IO at all.

TEST=unit,trybot,platform_CryptohomeTestAuth
BUG=chromium-os:20778

Change-Id: I202aafcc23342c249e255bf4b79cd01144d21a66
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19904
Reviewed-on: https://gerrit.chromium.org/gerrit/20398
Reviewed-by: Will Drewry <wad@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>