Fixed a race condition between chapsd and chaps clients.
This race condition exhibits itself in practice when cryptohome restarts
chapsd and then waits for the process to be respawned before making
calls. If cryptohome calls between the time that chapsd has started and
the time it begins servicing dbus then the calls fail.

This is suspected of causing apparent certificate loss.

BUG=chromium:120767
TEST=Ran all unit tests.
     Manually tested:
     - Ran client but not service.
     - Ran client and then service after a short delay.
     - Ran service and then client.

Change-Id: I0a9105e3d10b0b359ed3faa0c91516c57dffa304