chromeos_startup: change mount-encrypted flag file
In order to support using crosh to set a feature flag, this test
needs to become a two-stage check. If the system is running without
mount-encrypted, then /home/chronos exists in the clear for crosh to
operate on and for chromeos_startup to read. On a reboot, mount-encrypted
will migrate /home/chronos to inside the encrypted partition, making this
flag invisible to chromeos_startup, but it does create the encrypted.key
file. On a boot with mount-encrypted already running, the indication of
the need to run will be the key file itself.

Turning on mount-encrypted is basically a one-way operation, since there
is no "unmigrate" path, and turning it off is the same as wiping the
stateful partition. So, to turn it off, wipe stateful. :)

BUG=chromium-os:22172
TEST=x86-alex build, manual testing

Change-Id: If436544ca26b833c4c3455d929a21309974d7c1a
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20515
Reviewed-by: Elly Jones <ellyjones@chromium.org>