Add CHECKs to prove our Xrandr use is symmetric and consistent.
Remove an extraneous call to ClearXrandr in Run and add CHECKs to
prove that our calls to Setup/Clear Xrandr are balanced.
Also added CHECKs after our calls to GetModeInfo to ensure that we
succeed in finding the mode info since we expect that we should never
fail to find the mode structure in the screen when we found the mode ID
in one of the screen's outputs.  That is, we expect that any failure to
find the mode would need to be a bug elsewhere (potentially stale screen
resources).
The bug in issue chromium-os:29536 has no information for recreation but
we suspect that recent changes in this file has fixed an asymmetry
causing the failure.  If that is not the case, these CHECKs will
flush the bug out closer to its source and with more concrete
assumptions.

BUG=chromium-os:29536
TEST=Ran --test and manually confirmed mode switching works on a Lumpy
connected to a 30-inch HP monitor over DP

Change-Id: Ia91b63820a85a38b44500d3da25a98f3ed38bdda
Reviewed-on: https://gerrit.chromium.org/gerrit/20483
Commit-Ready: Jeff Disher <disher@chromium.org>
Tested-by: Jeff Disher <disher@chromium.org>
Reviewed-by: Jeff Disher <disher@chromium.org>