shill: dbus_properties: Convert to callbacks instead of delegates
Use callbacks instead of proxies on the DBusPropertiesProxyInterface.
This makes the class consistent with the other proxies used for
cellular support, and in subsequent CLs will allow access to the
GetAll function without having to implement the Delegate interface.

Add a GetStrings method to dbus_properties.

BUG=none
TEST=run unit tests, run on a machine validate that changes in signal
strength are noticed.

Change-Id: I23e47f011c39a23f406ef62a52a5a613e55c55ed
Reviewed-on: https://gerrit.chromium.org/gerrit/20640
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: Jason Glasgow <jglasgow@chromium.org>
Commit-Ready: Jason Glasgow <jglasgow@chromium.org>