Add in Makefile support to power_manager
This CL adds in a common.mk based Makefile that is designed to replicate the
behaviour of our current SCons based build system. Since it replicates the
behaviour, it also has the same/similar inefficiences wrt things like over
specified dependencies. I am intending to clean up the code in future CLs, but
this CL is meant to have a working replacement for the build infrastructure.

The Makefile builds everthing in src/platform/power_manager/out instead of
src/platform/power_manager. This is because common.mk builds lots of
intermediary files, which are a pain if you are doing a cros_workon_make. The
libraries also are named libblah.pic.a instead of libblah.a as they are in
SCons. This is because common.mk supports both PIC and PIE builds and the name
tells the build system which one to build. If there are issues in testing from
other packages depending on the lib names, I will add rules to rename them after
building.

This CL also moves xidle_example.cc out of its own directory. This was done
because I would have had to add a module.mk in the example directory to support
this layout. This didn't seem to make much sense to me, so I moved the file and
changed SConstruct to match.

BUG=chromium-os:29524

These tests assume that we are still using the old SCons builds. They are done
to confirm that I didn't break anything when moving xidle_example.cc
TEST=Built package using 'cros_work_make power_manager' and confirmed correct
     results.
TEST=Built package using 'cros_work_make power_manager --test' and confirmed
     correct results.
TEST=Built package using 'emerge-x86-alex power_manager' and confirmed correct
     results.

All of the following tests assume that the ebuild for the package has been
changed to us make instead of scons. I have a CL for this,
https://gerrit.chromium.org/gerrit/21246.
TEST=Built package using 'cros_work_make power_manager' and confirmed correct
     results.
TEST=Built package using 'cros_work_make power_manager --test' and confirmed
     correct results.
TEST=Built package using 'emerge-x86-alex power_manager' and confirmed correct
     results.
TEST=Built x86-alex image to confirm other packages did not break.
TEST=Ran x86-alex image to confirm expected behaviour.
TEST=Built stumpy image to confirm other packages did not break.
TEST=Ran stumpy image to confirm expected behaviour.

Change-Id: I3d368b7f362967b8e893098e89cdd6b6f960f123
Reviewed-on: https://gerrit.chromium.org/gerrit/21245
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris Wolfe <cwolfe@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>