backlight controller: Check adjustment intervals
Adding a check to the brightness adjustment loop to make sure the actual
adjustment step intervals are close to the expected step intervals.
The amount of tolerance is +/-20%.

This patch also refactors the adjustment loop so that there is one
repeating event source for a series of adjustment steps instead of one
source for each step.

BUG=chrome-os-partner:8310
TEST=backlight controller unit test passes, gradual backlight
transitions work.

Change-Id: I55e425ef3d3eec4c30f6bd3b0322e3b4d27ee8d0
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20639