shill: Remove unnecessary extern "C" declaration around glib-unix.h
In glib 2.30, there is a missing extern "C" declaration in glib-unix.h.
An extern "C" declaration was added around the inclusion of glib-unix.h
in shill to address that issue. This is no longer needed as the
dev-libs/glib-2.30.0-r2 package has a patch that fixes the issue.

BUG=none
TEST=Build and run unit tests.

Change-Id: I69927d5e5416966c8feee8f17587ce9a11ccf873
Reviewed-on: https://gerrit.chromium.org/gerrit/20669
Tested-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Ben Chan <benchan@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>