Replace glog/logging with base/logging
glog/logging causes data to be written out to the backing store every
time a log message is sent to it.  This causes increased power draw
since the disk wakes up more frequently than it would if the data were
to be written out lazily.

Replacing glog/logging with base/logging since the latter doesn't have
this issue.

BUG=chromium-os:29643
TEST="start cromo" on DUT and make sure it comes up fine

Change-Id: I2862ad9da70754112dc3c2fbe0714bd964127872