debugd: write null clock sync record
Write a clock sync record on systrace stop that causes chrome to not
time-shift data.  This used in conjunction with other changes that export
the kernel's trace event clock to user space to align user and kernel
event data.

With this change we no longer need the clock_monotonic helper so delete it.

BUG=chromium-os:27809
TEST=manual:collect trace data; check sync record contents

Change-Id: I2d7edcc9e6b80bee96033ec00b4dddef016f8307