power: Create BacklightController interface.
This renames the existing BacklightController class to
InternalBacklightController and adds a new
BacklightController interface.

In a later change, I plan to add a new
ExternalBacklightController class for devices that lack
internal displays.  Its behavior will differ from that of
InternalBacklightController (in general, it'll be simpler --
the backlight level will only be changed in response to the
brightness keys, and it'll emit signals asking Chrome to dim
the screen just before we suspend).

BUG=chromium-os:27501,chromium-os:28663,chromium:125176
TEST=existing tests pass on stumpy and lumpy

Change-Id: I69e3ae02ea44eac9c0ad8f814b649867e7a77551
Reviewed-on: https://gerrit.chromium.org/gerrit/21269
Reviewed-by: Simon Que <sque@chromium.org>
Commit-Ready: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>