power: Create ExternalBacklightController.
This creates a new BacklightController implementation for
use on desktop machines.  The class makes no automated
changes to the external display's brightness level, only
adjusting it in response to the brightness keys being
pressed.  When we enter the idle state in response to user
inactivity, it emits a D-Bus signal asking Chrome to dim the
screen in software.

BUG=chromium-os:27501,chromium-os:28663,chromium:125176
TEST=added; also did manual testing

Change-Id: Iecfff1eb28145c3be3bc26a0d3141a643d0026fe
Reviewed-on: https://gerrit.chromium.org/gerrit/21466
Commit-Ready: Daniel Erat <derat@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>