Make RollingAverage more resiliant to error conditions.
Many of the error conditions in RollingAverage, which is used by powerd, are
currently fatal since they are tested using CHECKs. This means that as new
platforms are being brought up powerd is very crashy, which is a pain point for
developers. These conditions can be recovered from normally or are not severe
enough to justify killing powerd, since they will only result in weird values in
a UX dialogue.

This CL removes all of the CHECK statements from RollingAverage and replaces
them with appropriate logging messages and error handling. It also updates the
associated unit tests.

BUG=chromium-os:29967
TEST=Built package and ran unittests
TEST=Built image and confirmed that no regressions were introduced
TEST=Ran trybot for amd64-generic-paladin

Change-Id: I0394d2b9ee143c1cd2850a25285b3652dc397bd4
Reviewed-on: https://gerrit.chromium.org/gerrit/21334
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>