Remove SConstruct
Since we are moving from SCons to Make this file is no longer needed. Removing
it, so that people don't make changes to it and expect it to affect the build.

The two dependencies for this CL are the CLs that will add Makefile support and
turn it on for power_manager.

CQ-DEPEND=I68bbf9692410c774e957872fe61a44dce0211fd2
CQ-DEPEND=I3d368b7f362967b8e893098e89cdd6b6f960f123
BUG=chromium-os:29524
TEST=Built power_manager package

Change-Id: Ic70620e1a922b191ee3b3b7d2023d93ffca2f85b
Reviewed-on: https://gerrit.chromium.org/gerrit/21528
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>