power: add workaround for reviving gobi modem on resume
On some Lumpy systems the gobi modem goes missing in action after
resuming.  The workaround is to toggle the power and off GPIOs going to
the gobi modem.

Adding workaround on the resume path to toggle these GPIOs if the gobi
modem was present prior to suspend but is missing coming out of resume.

During testing I am finding that the gobi modem is never found on the
resume path in powerd_suspend since the USB resume happens
asynchronously and isn't done by the time kernel hands back control to
userland on resume. This means that the workaround code is always
executed on resume path.

BUG=chrome-os-partner:9320
TEST=suspend the system using powerd_suspend or closing lid etc. After
the system has resumed run "modem status" command and check that the
Gobi modem is listed.

Change-Id: I78df6f48d346595f8faee60f919806819d2e45bb
Reviewed-on: https://gerrit.chromium.org/gerrit/21835
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Ready: Sameer Nanda <snanda@chromium.org>
Tested-by: Sameer Nanda <snanda@chromium.org>